move day_time test
